[BANK-6021] Add Signing Functionality (#422)
This PR adds signing capabilities to Sample App which is necessary to
complete a StableFX trade in testing.
The provider for signing we use is Circle Developer Controlled Wallets;
New settings are added to the configuration drawer to input a API key,
entitysecret, and walletId for signing requests.
StableFX Presign API pages are updated to be able to sign typed data
after retrieving a response.

See below for a demo,


https://github.com/user-attachments/assets/28cbb424-46ea-4921-977d-11f5ac2366e5